// src/redux/types.ts
export enum types {
  SIGN_UP_REQUESTED = "auth/SIGN_UP_REQUESTED",
  SIGN_IN_UP = "auth/SIGN_IN_UP",
  SIGN_IN_REQUESTED = "auth/SIGN_IN_REQUESTED",
  SET_IS_LOGGED_IN = "auth/SET_IS_LOGGED_IN",
  SET_TOKEN = "auth/SET_TOKEN",
  SET_TOKEN_REQUESTED = "auth/SET_TOKEN_REQUESTED",
  LOG_OUT = "auth/LOG_OUT",
  LOG_OUT_REQUESTED = "auth/LOG_OUT_REQUESTED",

  SET_COLUMNS = "columns/SET_COLUMNS",
  FETCH_COLUMNS = "columns/FETCH_COLUMNS",
  CREATE_COLUMN = "column/CREATE_COLUMN",
  CREATE_COLUMN_REQUESTED = "column/CREATE_COLUMN_REQUESTED",
  UPDATE_COLUMN = "column/UPDATE_COLUMN",
  UPDATE_COLUMN_REQUESTED = "column/UPDATE_COLUMN_REQUESTED",
  DELETE_COLUMN = "column/DELETE_COLUMN",
  DELETE_COLUMN_REQUESTED = "column/DELETE_COLUMN_REQUESTED",
  GET_COLUMN = "column/GET_COLUMN",
  FETCH_COLUMN = "column/FETCH_COLUMN",

  SET_PRAYERS = "prayers/SET_PRAYERS",
  FETCH_PRAYERS = "prayers/FETCH_PRAYERS",
  GET_PRAYER = "prayer/GET_PRAYER",
  FETCH_PRAYER = "prayer/FETCH_PRAYER",
  DELETE_PRAYER = "prayers/DELETE_PRAYER",
  DELETE_PRAYER_REQUESTED = "prayers/DELETE_PRAYER_REQUESTED",
  CREATE_PRAYER = "prayers/CREATE_PRAYER",
  CREATE_PRAYER_REQUESTED = "prayers/CREATE_PRAYER_REQUESTED",
  UPDATE_PRAYER = "prayers/UPDATE_PRAYER",
  UPDATE_PRAYER_REQUESTED = "prayers/UPDATE_PRAYER_REQUESTED",

  SET_COMMENTS = "comments/SET_COMMENTS",
  FETCH_COMMENTS = "comments/FETCH_COMMENTS",
  CREATE_COMMENT = "comments/CREATE_COMMENT",
  CREATE_COMMENT_REQUESTED = "comments/CREATE_COMMENT_REQUESTED",
  DELETE_COMMENT = "comments/DELETE_COMMENT",
  DELETE_COMMENT_REQUESTED = "comments/DELETE_COMMENT_REQUESTED",

  APP_SET_STATUS = "app/SET-STATUS"
}